# api/openapi.yaml
openapi: 3.0.3
info:
  title: Bilibili RAG 知识库系统 API
  description: |
    B站收藏内容可检索知识库系统的 RESTful API 文档。

    ## 核心能力

    1. **数据获取** - B站扫码登录、收藏夹/视频列表同步
    2. **内容提取** - ASR（语音转文本）、基础信息（标题/简介）
    3. **知识构建** - 文本切块（chunking）、向量化（embedding）、存储到 ChromaDB
    4. **智能问答（RAG）** - 查询路由（direct/db/vector）、向量检索+LLM生成、流式返回+来源追踪

    ## 认证机制

    - 使用 `session_id` 作为会话标识
    - 扫码登录后获取 `session_id`
    - 大部分接口需要携带 `session_id` 参数

    ## SSE 流式响应

    聊天接口 `/chat/ask/stream` 使用 Server-Sent Events 流式返回：
    - 格式: `data: {chunk_text}`
    - 结束后发送: `data: [[SOURCES_JSON]]{sources_json}`
  version: 1.2.0
  contact:
    name: Bilibili RAG Project
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
...
paths:
...
  /asr/content:
    get:
      tags:
        - ASR
      summary: 查询 ASR 内容
      description: |
        查询指定分P的 ASR 转写内容。

        - 不存在返回 `{exists: false}`
        - 存在返回内容详情
      operationId: getASRContent
      parameters:
        - name: bvid
          in: query
          required: true
          description: 视频BV号
          schema:
            type: string
        - name: cid
          in: query
          required: true
          description: 分P的CID
          schema:
            type: integer
      responses:
        '200':
          description: ASR内容
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ASRContentResponse'
              example:
                exists: true
                bvid: "BV1xx411c7mD"
                cid: 12345678
                page_index: 0
                page_title: "P1: 课程介绍"
                content: "欢迎来到本课程..."
                content_source: "asr"
                version: 1
                is_processed: true
  /asr/create:
    post:
      tags:
        - ASR
      summary: 创建 ASR 任务
      description: |
        幂等创建 ASR 转写任务。

        - 已存在且 `is_processed=true` → 直接返回
        - 不存在 → 创建记录并启动后台任务
      operationId: createASR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ASRCreateRequest'
            example:
              bvid: "BV1xx411c7mD"
              cid: 12345678
              page_index: 0
              page_title: "P1: 课程介绍"
      responses:
        '200':
          description: 任务创建结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    nullable: true
                  message:
                    type: string
                  version:
                    type: integer
              example:
                task_id: "550e8400-e29b-41d4-a716-446655440000"
                message: "ASR 任务已创建"
        '409':
          description: ASR 已完成，无需重复创建
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                    nullable: true
                  message:
                    type: string
  /asr/update:
    post:
      tags:
        - ASR
      summary: 更新 ASR 内容
      description: |
        手动编辑并覆盖 ASR 转写内容。不新建版本，直接覆盖当前内容。
      operationId: updateASR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ASRUpdateRequest'
            example:
              bvid: "BV1xx411c7mD"
              cid: 12345678
              page_index: 0
              content: "手动编辑后的内容..."
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
              example:
                success: true
                message: "更新成功"
        '404':
          description: ASR 记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /asr/reasr:
    post:
      tags:
        - ASR
      summary: 重新 ASR
      description: |
        强制重新执行 ASR 转写。

        - 将当前内容保存到版本历史
        - 清空当前内容，启动新的 ASR 任务
      operationId: reASR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ASRReASRRequest'
            example:
              bvid: "BV1xx411c7mD"
              cid: 12345678
              page_index: 0
      responses:
        '200':
          description: 重新 ASR 已启动
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    type: string
                  message:
                    type: string
              example:
                task_id: "550e8400-e29b-41d4-a716-446655440000"
                message: "重新 ASR 已启动"
        '404':
          description: ASR 记录不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /asr/status/{task_id}:
    get:
      tags:
        - ASR
      summary: 查询 ASR 任务状态
      description: 轮询 ASR 后台任务的执行状态。
      operationId: getASRTaskStatus
      parameters:
        - name: task_id
          in: path
          required: true
          description: 任务ID
          schema:
            type: string
      responses:
        '200':
          description: 任务状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ASRTaskStatus'
              example:
                task_id: "550e8400-e29b-41d4-a716-446655440000"
                status: "processing"
                progress: 65
                message: "转写中..."
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /asr/versions:
    get:
      tags:
        - ASR
      summary: 查询 ASR 版本历史
      description: 获取指定分P的 ASR 版本历史列表。
      operationId: getASRVersions
      parameters:
        - name: bvid
          in: query
          required: true
          description: 视频BV号
          schema:
            type: string
        - name: cid
          in: query
          required: true
          description: 分P的CID
          schema:
            type: integer
      responses:
        '200':
          description: 版本历史列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VideoPageVersionInfo'
              example:
                - version: 1
                  content_source: "asr"
                  content_preview: "欢迎来到本课程..."
                  is_latest: false
                  created_at: "2026-04-01T10:00:00"
                - version: 2
                  content_source: "asr"
                  content_preview: "欢迎来到本课程（修正版）..."
                  is_latest: true
                  created_at: "2026-04-02T10:00:00"

  # ========== 分P向量化 ==========
...
components:
  schemas:
...
    ASRCreateRequest:
      type: object
      required:
        - bvid
        - cid
      properties:
        bvid:
          type: string
          description: 视频BV号
        cid:
          type: integer
          description: 分P的CID
        page_index:
          type: integer
          description: P序号（0-based）
          default: 0
        page_title:
          type: string
          description: 分P标题
          nullable: true
    ASRUpdateRequest:
      type: object
      required:
        - bvid
        - cid
        - page_index
        - content
      properties:
        bvid:
          type: string
        cid:
          type: integer
        page_index:
          type: integer
        content:
          type: string
          description: 手动编辑的ASR内容
    ASRReASRRequest:
      type: object
      required:
        - bvid
        - cid
        - page_index
      properties:
        bvid:
          type: string
        cid:
          type: integer
        page_index:
          type: integer
    ASRContentResponse:
      type: object
      required:
        - exists
      properties:
        exists:
          type: boolean
          description: 是否存在ASR记录
        bvid:
          type: string
          nullable: true
        cid:
          type: integer
          nullable: true
        page_index:
          type: integer
          nullable: true
        page_title:
          type: string
          nullable: true
        content:
          type: string
          description: ASR转写全文
          nullable: true
        content_source:
          type: string
          description: 内容来源（asr / user_edit）
          nullable: true
        version:
          type: integer
          nullable: true
        is_processed:
          type: boolean
          nullable: true
    ASRTaskStatus:
      type: object
      required:
        - task_id
        - status
        - progress
        - message
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [pending, processing, done, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
    VideoPageVersionInfo:
      type: object
      required:
        - version
        - content_source
        - content_preview
        - is_latest
      properties:
        version:
          type: integer
        content_source:
          type: string
        content_preview:
          type: string
        is_latest:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ========== 分P向量化 ==========
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: 错误详情